# api/openapi.yaml
openapi: 3.1.0
info:
  title: Lux API
  version: v1
  description: "The control plane of Lux under /v1: apply a Provider, Model, Key, or Budget by name with PUT, read one by id or name, list with limit and cursor, delete, and rotate a Key. Every request body is a manifest of lux.latere.ai/v1beta1 in JSON or YAML; every response is JSON; every refusal is one envelope with one code from the error table in x-lux-errors. Every URL is under LUX_PUBLIC_URL."
paths:
...
  /v1/usage:
    get:
      operationId: readUsage
      summary: Aggregate usage
      description: Usage aggregated over a range, grouped by at most three dimensions and bucketed by an interval; unpaged, and no row sums two currencies. A filter outside the authorizer's own is an empty items.
      x-lux-action: usage.read
      parameters:
      - $ref: "#/components/parameters/from"
      - $ref: "#/components/parameters/to"
      - $ref: "#/components/parameters/by"
      - $ref: "#/components/parameters/interval"
      - $ref: "#/components/parameters/usageKey"
      - $ref: "#/components/parameters/usageModel"
      - $ref: "#/components/parameters/usageProvider"
      - $ref: "#/components/parameters/usageOwner"
      - $ref: "#/components/parameters/usageLabel"
      responses:
        "200":
          description: The rows.
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageList"
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/usage/redact:
    post:
      operationId: redactUsage
      summary: Redact usage owner
      description: Every hourly and monthly usage row of the owner, and every record of this replica's ring, loses the owner; each row's sums are added into the row with the same other dimensions and no owner, so totals are kept.
      x-lux-action: usage.redact
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageRedaction"
      responses:
        "200":
          description: The rows rewritten.
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageRedacted"
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/requests:
    get:
      operationId: listRequests
      summary: List requests
      description: Usage records over a range, newest first, paged by limit and cursor, with the record set that answered beside them.
      x-lux-action: usage.read
      parameters:
      - $ref: "#/components/parameters/from"
      - $ref: "#/components/parameters/to"
      - $ref: "#/components/parameters/usageKey"
      - $ref: "#/components/parameters/usageModel"
      - $ref: "#/components/parameters/usageProvider"
      - $ref: "#/components/parameters/usageOwner"
      - $ref: "#/components/parameters/usageLabel"
      - $ref: "#/components/parameters/status"
      - $ref: "#/components/parameters/error"
      - $ref: "#/components/parameters/stream"
      - $ref: "#/components/parameters/recordLimit"
      - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: One page of records.
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordList"
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/self:
    get:
      operationId: readSelf
      summary: Read caller identity
      description: The caller's identity, who decides permission, and what this replica remembers granting the subject.
      x-lux-action: none
      responses:
        "200":
          description: The caller.
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Self"
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/openapi.json:
    get:
      operationId: readOpenAPI
      summary: Read OpenAPI document
      description: This document as JSON; no bearer.
      x-lux-action: none
      parameters: []
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema:
                type: object
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    security: []
  /.well-known/lux:
    get:
      operationId: readWellKnown
      summary: Read server identity
      description: "The server's identity: the build, the API and the doors under LUX_PUBLIC_URL, the issuers, the audience, and the mode; no bearer."
      x-lux-action: none
      parameters: []
      responses:
        "200":
          description: The server.
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WellKnown"
        default:
          description: "A refusal: one code of x-lux-errors with its fixed sentence."
          headers:
            Lux-Request-Id:
              $ref: "#/components/headers/Lux-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A token from an issuer in LUX_OIDC_ISSUERS with aud containing LUX_OIDC_AUDIENCE.
  parameters:
...
    cursor:
      name: cursor
      in: query
      required: false
      description: The previous page's next_cursor, opaque; one from another kind or filter is invalid_field.
      schema:
        type: string
...
    from:
      name: from
      in: query
      required: false
      description: The start of the range, RFC 3339; default 24 hours before to. The range is at most 90 days.
      schema:
        type: string
        format: date-time
    to:
      name: to
      in: query
      required: false
      description: The end of the range, RFC 3339; default now. It must be after from.
      schema:
        type: string
        format: date-time
    by:
      name: by
      in: query
      required: false
      description: "The dimensions to group by, comma separated, at most three: key, model, provider, owner, door, status, or label:<name>. None is one total row."
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    interval:
      name: interval
      in: query
      required: false
      description: The bucket width; none is one row per group over the whole range.
      schema:
        type: string
        enum:
        - none
        - hour
        - day
        - month
        default: none
    usageKey:
      name: key
      in: query
      required: false
      description: A Key by name or key_ id, repeatable; a filter, not a grouping. A name that names no Key is an empty items, and a key_ id keeps working after the Key is deleted.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    usageModel:
      name: model
      in: query
      required: false
      description: A Model by name or mdl_ id, repeatable.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    usageProvider:
      name: provider
      in: query
      required: false
      description: A Provider by name or prv_ id, repeatable.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    usageOwner:
      name: owner
      in: query
      required: false
      description: A rendered subject, repeatable; intersected with the authorizer's filter, so one outside it is an empty items.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    usageLabel:
      name: label
      in: query
      required: false
      description: name=value over the Key's labels, repeatable; every pair must match.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    status:
      name: status
      in: query
      required: false
      description: Records of one outcome.
      schema:
        type: string
        enum:
        - ok
        - refused
        - failed
    error:
      name: error
      in: query
      required: false
      description: Records carrying one error code of x-lux-errors.
      schema:
        type: string
    stream:
      name: stream
      in: query
      required: false
      description: Streamed records alone, or unstreamed alone.
      schema:
        type: boolean
    recordLimit:
      name: limit
      in: query
      required: false
      description: Page size, default 50, at most 1000; above is invalid_field.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
...
  headers:
    Lux-Request-Id:
      description: The server's id of this request, req_ and a ULID; a caller's own is replaced.
      schema:
        type: string
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              enum:
              - malformed_body
              - multi_document
              - unsupported_version
              - unsupported_kind
              - unknown_field
              - missing_field
              - invalid_field
              - reserved_prefix
              - exclusive_fields
              - duplicate_target
              - invalid_request
              - upstream_rejected
              - dialect_unsupported
              - provider_required
              - currency_mismatch
              - unauthenticated
              - forbidden
              - key_disabled
              - key_expired
              - route_not_allowed
              - model_not_allowed
              - model_unpriced
              - not_found
              - model_not_found
              - read_only
              - already_exists
              - key_fenced
              - fence_conflict
              - conflict
              - immutable_field
              - budget_in_use
              - provider_in_use
              - body_too_large
              - unsupported_media_type
              - ceiling_exceeded
              - rate_limited
              - spend_exceeded
              - budget_exhausted
              - internal
              - upstream_error
              - authorizer_unavailable
              - store_unavailable
              - provider_unavailable
              - upstream_timeout
            message:
              type: string
              description: The one fixed user sentence of the code.
            details:
              type: object
              required:
              - request_id
              properties:
                request_id:
                  type: string
                paths:
                  type: array
                  items:
                    type: string
                  description: The JSON paths of the fields the code names.
                detail:
                  type: string
                  description: One developer sentence.
...
    UsageRedaction:
      type: object
      additionalProperties: false
      required:
      - owner
      properties:
        owner:
          type: string
          description: Rendered issuer|subject whose identity every usage row and ring record loses.
    UsageRedacted:
      type: object
      required:
      - rows
      properties:
        rows:
          type: integer
          description: The usage rows rewritten.
...
    Self:
      $ref: "#/components/schemas/Self"
...
    WellKnown:
      $ref: "#/components/schemas/WellKnown"
    Row:
      type: object
      required:
      - dimensions
      - requests
      - ok
      - refused
      - failed
      - inputTokens
      - outputTokens
      - cachedInputTokens
      - cacheWriteTokens
      - cost
      - currency
      - unpricedRequests
      properties:
        bucket:
          type: string
          format: date-time
        dimensions:
          type: object
          additionalProperties:
            type: string
        requests:
          type: integer
        ok:
          type: integer
        refused:
          type: integer
        failed:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        cachedInputTokens:
          type: integer
        cacheWriteTokens:
          type: integer
        cost:
          type: integer
        currency:
          type: string
        unpricedRequests:
          type: integer
    UsageList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Row"
    Record:
      type: object
      required:
      - id
      - at
      - endedAt
      - key
      - owner
      - model
      - provider
      - upstreamModel
      - door
      - targetDialect
      - route
      - translated
      - loss
      - attempts
      - status
      - error
      - upstreamStatus
      - latencyMs
      - ttfbMs
      - tokens
      - cost
      - stream
      - labels
      - requestLabels
      properties:
        id:
          type: string
        at:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        key:
          $ref: "#/components/schemas/KeyRef"
        owner:
          type: string
        model:
          $ref: "#/components/schemas/Ref"
        provider:
          $ref: "#/components/schemas/Ref"
        upstreamModel:
          type: string
        door:
          type: string
          enum:
          - openai
          - anthropic
          - gemini
          - lux
        targetDialect:
          type: string
          enum:
          - ""
          - openai
          - anthropic
          - gemini
          - lux
          description: The dialect of the target that answered; empty when no target was chosen, as on a refused request.
        route:
          type: string
        translated:
          type: boolean
        loss:
          type: array
          items:
            type: string
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/Attempt"
        status:
          type: string
          enum:
          - ok
          - refused
          - failed
        error:
          type: string
        upstreamStatus:
          type: integer
        latencyMs:
          type: integer
        ttfbMs:
          type: integer
        tokens:
          $ref: "#/components/schemas/Tokens"
        cost:
          $ref: "#/components/schemas/Charge"
        stream:
          type: boolean
        labels:
          type: object
          additionalProperties:
            type: string
        requestLabels:
          type: object
          additionalProperties:
            type: string
    KeyRef:
      type: object
      required:
      - id
      - prefix
      properties:
        id:
          type: string
        prefix:
          type: string
    Ref:
      type: object
      required:
      - name
      - id
      properties:
        name:
          type: string
        id:
          type: string
    Attempt:
      type: object
      required:
      - provider
      - upstreamModel
      - status
      - httpStatus
      - error
      - durationMs
      properties:
        provider:
          type: string
        upstreamModel:
          type: string
        status:
          type: string
          enum:
          - ok
          - refused
          - failed
        httpStatus:
          type: integer
        error:
          type: string
        durationMs:
          type: integer
    Tokens:
      type: object
      required:
      - input
      - output
      - cachedInput
      - cacheWrite
      - reasoning
      - estimated
      properties:
        input:
          type: integer
        output:
          type: integer
        cachedInput:
          type: integer
        cacheWrite:
          type: integer
        reasoning:
          type: integer
        estimated:
          type: boolean
    Charge:
      type: object
      required:
      - amount
      - currency
      - priced
      properties:
        amount:
          type: integer
        currency:
          type: string
        priced:
          type: boolean
    RecordList:
      type: object
      required:
      - items
      - source
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Record"
        next_cursor:
          type: string
          description: The cursor of the next page; absent on the last.
        source:
          type: string
          enum:
          - archive
          - memory
          description: "Which record set answered: the replica's own ring, or the installation-wide archive when a request log exporter is configured."
security:
- bearer: []
...